pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "texword"
version = "0.1.0"
description = "High-quality academic paper converter: LaTeX/PDF → editable Word"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Yang Zhengyuan"},
]
keywords = ["latex", "word", "docx", "academic", "converter", "pandoc"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Text Processing :: Markup :: LaTeX",
]
dependencies = [
    "python-docx>=0.8.11",
    "PyMuPDF>=1.23.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
texword = "texword.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["texword*"]
